requirements.txt
pyperclip==1.8.2
requests==2.30.0
bs4==0.0.2
beautifulsoup4==4.12.3
lxml==5.1.0
Pillow==9.5.0
watchdog==3.0.0

numpy~=1.26.4
opencv-python~=4.9.0.80
selenium~=4.20.0
pygame~=2.5.2
imageio~=2.34.0
ephem~=4.1.5
ninepatch~=0.2.0
pyzipper~=0.3.6